fix:count의 타입 Long으로 수정 및 타입차이 cast로해결

+++ src/main/java/com/danggeun/market/interest/service/InterestProductSearchService.java
@@ -30,8 +30,8 @@ public class InterestProductSearchService {
         List<Object[]> result = interestHistoryRepository.searchInterestProduct(user);
         return result.stream()
                 .map(objects -> new ProductSummaryResponse(((InterestHistory) objects[0]).getProduct(),
-                        (Integer) objects[1],
-                        (Integer) objects[2]))
+                        (Long) objects[1],
+                        (Long) objects[2]))
                 .collect(Collectors.toList());
     }
 }
